Revert "Mörkö SFX and soundsources"
