--- Assets/MorkoSFXController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02a0c867dc075de4fbedb6925cf80c36
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
