--- Assets/Prefabs/SoundEvents/SFXPlayCounter.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: eac1211afb906354fb9809acbb43cae5
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
