+++ Assets/Prefabs/SoundEvents/SoundSourcePrefabs/AdjustableSoundSource.prefab
@@ -105,7 +105,7 @@ AudioSource:
     m_Curve:
     - serializedVersion: 3
       time: 0
-      value: 1
+      value: 0.8
       inSlope: 0
       outSlope: 0
       tangentMode: 0
