+++ Assets/Prefabs/SoundEvents/TriggerPrefabs/AudioTriggersFirstHalf.prefab
@@ -1,96 +1,5 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &582145356
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 582145357}
-  - component: {fileID: 582145359}
-  - component: {fileID: 582145358}
-  m_Layer: 0
-  m_Name: MusicFadeOutAndInTrigger
-  m_TagString: Untagged
-  m_Icon: {fileID: 7250588514170254948, guid: 0000000000000000d000000000000000, type: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!4 &582145357
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 582145356}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 272.27, y: -9.79, z: 0}
-  m_LocalScale: {x: 0.1, y: 8, z: 1}
-  m_Children: []
-  m_Father: {fileID: 2844684736883143559}
-  m_RootOrder: 20
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!61 &582145359
-BoxCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 582145356}
-  m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0, y: 0}
-    oldSize: {x: 0, y: 0}
-    newSize: {x: 0, y: 0}
-    adaptiveTilingThreshold: 0
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 1, y: 1}
-  m_EdgeRadius: 0
---- !u!114 &582145358
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 582145356}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e65429390e67e9b4098c599c7f97028b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  destroyableSoundSource: {fileID: 0}
-  audioOptions: 8
-  _music: {fileID: 0}
-  _sfx: {fileID: 0}
-  _musicToFadeOut: {fileID: 0}
-  _musicToFadeIn: {fileID: 0}
-  _sfxToFadeOut: {fileID: 0}
-  _sfxToFadeIn: {fileID: 0}
-  audioSnapshot: []
-  mixer: {fileID: 0}
-  weights: []
-  loop: 0
-  destroyOnTrigger: 0
-  musicFadeTime: 5
-  sfxFadeTime: 4
-  volume: 0
-  desiredIncreasedMusicVolume: 1
-  desiredIncreasedSFXVolume: 0
-  desiredFadedMusicVolume: 0
-  desiredFadedSFXVolume: 0
-  desiredPitch: 1
 --- !u!1 &2844684735918605146
 GameObject:
   m_ObjectHideFlags: 0
@@ -829,7 +738,7 @@ GameObject:
   m_Component:
   - component: {fileID: 2844684736883143559}
   m_Layer: 0
-  m_Name: AudioTriggersFirstHalf
+  m_Name: AudioTriggersFirstHalf 1
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -866,7 +775,6 @@ Transform:
   - {fileID: 6075131511856452841}
   - {fileID: 3598345469948177961}
   - {fileID: 3598345470169032354}
-  - {fileID: 582145357}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
