+++ Assets/Scripts/AudioEffects/SFXIntervalandPitchRandomizer.cs
@@ -6,32 +6,29 @@ public class SFXIntervalandPitchRandomizer : MonoBehaviour
 {
     public AudioSource _audio;
 
-    private SFXPlayCounter counter;
+    private float counter;
 
     public int possibleStartTime;
     public int possibleEndTime;
     public float pitchLowerBorder;
     public float pitchHigherBorder;
 
-    private int sum;
-
     private int rand;
 
     private void Start()
     {
-        counter = GameObject.Find("SFXPlayCounter").GetComponent<SFXPlayCounter>();
         rand = Random.Range(possibleStartTime, possibleEndTime);
-        sum = rand;
     }
     // Update is called once per frame
     void FixedUpdate()
     {
-        if (counter.counter >= sum)
+        counter += Time.deltaTime;
+        if (counter >= rand)
         {
             _audio.pitch = (float)Random.Range(pitchLowerBorder, pitchHigherBorder);
             _audio.Play();
+            counter = 0;
             rand = Random.Range(possibleStartTime, possibleEndTime);
-            sum += rand;
         }
     }
 }
