--- Assets/Scripts/AudioEffects/SFXPlayCounter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6182fcc99c678f4ca23b9eaef4816eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
