--- Assets/Scripts/Enemies/MorkoWalkAnimationCooldown.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f02b22c38eb84542be8e5ce88b7b755
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
